status codes

+++ Backend/server/src/main/java/com/uaic/server/controller/EmployeeController.java
@@ -3,7 +3,11 @@ package com.uaic.server.controller;
 import com.uaic.server.entities.Employee;
 import com.uaic.server.services.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/employees")
@@ -13,29 +17,45 @@ public class EmployeeController {
     private EmployeeService employeeService;
 
     @GetMapping
-    public Iterable<Employee> getEmployees() {
-        return employeeService.findEmployees();
+    public ResponseEntity<Iterable<Employee>> getEmployees() {
+        Iterable<Employee> employees = employeeService.findEmployees();
+        return new ResponseEntity<>(employees, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public Employee getEmployee(@PathVariable Integer id) {
-        return employeeService.findEmployeeById(id).isPresent()? employeeService.findEmployeeById(id).get(): null;
+    public ResponseEntity<Employee> getEmployee(@PathVariable Integer id) {
+        Optional<Employee> employee = employeeService.findEmployeeById(id);
+        if (employee.isPresent()) {
+            return new ResponseEntity<>(employee.get(), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
-    @PostMapping("{id}")
-    public void addEmployee(@PathVariable Integer id, @RequestBody Employee employee) {
-        employee.setId(id);
+    @PostMapping
+    public ResponseEntity<Void> addEmployee(@RequestBody Employee employee) {
         employeeService.createOrUpdateEmployee(employee);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public void updateEmployee(@PathVariable Integer id, @RequestBody Employee employee) {
-        employee.setId(id);
-        employeeService.createOrUpdateEmployee(employee);
+    public ResponseEntity<Void> updateEmployee(@PathVariable Integer id, @RequestBody Employee employee) {
+        if (employeeService.checkExistentEmployeeById(id)) {
+            employee.setId(id);
+            employeeService.createOrUpdateEmployee(employee);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteEmployee(@PathVariable Integer id) {
-        employeeService.deleteEmployeeById(id);
+    public ResponseEntity<Void> deleteEmployee(@PathVariable Integer id) {
+        if (employeeService.checkExistentEmployeeById(id)) {
+            employeeService.deleteEmployeeById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
